+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Define borehole csv import fields `original_name`, `location_x` and `location_y` as required.
+- Add validation for required borehole csv import fields. 
 - Added online help.
 
 ### Changed